requirements.txt
aiohttp==3.6.2
async-cse==0.3.0
async-timeout==3.0.1
DateTime==4.3
discord-ext-ipc @ git+git://github.com/ext-creators/discord-ext-ipc@2aaaa7b75bae0cb78888dc022585493a82c7f267
discord-ext-menus @ git+https://github.com/Rapptz/discord-ext-menus@84caae8038d0d3adc860957ccef05baeec2e2dd8
discord.py @ git+https://github.com/Rapptz/discord.py/@6515f339780ba2efe8951b827de53656aa84b695
aiowiki @ git+https://github.com/Gelbpunkt/aiowiki@2c342e748b63695413b56b7d04cdd6b2c9abfbf1
googletrans==3.1.0a0
humanize==3.1.0
jishaku==1.19.2.202
lavalink==3.1.2
motor==2.3.0
Pillow==8.0.1
psutil==5.7.3
Quart==0.14.1
tornado==6.0.4
twemoji-parser==0.4.2
wavelink==0.9.6
yarl==1.4.2
